allow disabling repsonse verification
I don't know why exporter_exporter tries to reason about the metrics in
the first place. Prometheus is able to decide itself what it can
eat or not.

Also the verification library is pinned to version 0.2.0. The current
one is 0.31.1 <https://github.com/prometheus/common/tree/v0.31.1/expfmt>

So as a bare minimum I want to be able to disable the verifcation on
a per module basis. I wouldn't bother if it is removed completely.
So this commit adds a key "no_validate_metrics" to the module config to
accomplish this.

Some background info:

We were trying to proxy to a jmx_exporter and it spit out the following
line:

    # TYPE jvm info

And exporter_exporter reacted with an internal server error and
complained about not knowing the type "info".